+++ EssentialFeedIITests/Feed Cache/CacheFeedImageDataUseCaseTests.swift
@@ -1,0 +1,31 @@
+//
+//  CacheFeedImageDataUseCaseTests.swift
+//  EssentialFeedIITests
+//
+//  Created by Julius on 25/10/2023.
+//
+
+import XCTest
+import EssentialFeedII
+
+final class CacheFeedImageDataUseCaseTests: XCTestCase {
+    func test_saveImageDataForURL_requestsImageDataInsertionForURL() {
+        let (sut, store) = makeSUT()
+        let data = anyData()
+        let url = anyURL()
+        
+        sut.save(data, for: url) { _ in }
+        
+        XCTAssertEqual(store.receivedMessages, [.insert(data: data, for: url)])
+    }
+    
+    // MARK: - Helpers
+    
+    private func makeSUT(file: StaticString = #file, line: UInt = #line) -> (sut: LocalFeedImageDataLoader, store: FeedImageDataStoreSpy) {
+        let store = FeedImageDataStoreSpy()
+        let sut = LocalFeedImageDataLoader(store: store)
+        trackForMemoryLeaks(store, file: file, line: line)
+        trackForMemoryLeaks(sut, file: file, line: line )
+        return (sut, store)
+    }
+}
